MSVC: Fix calls to pthread_mutexattr_init
The chosen implementation on MSVC does not accept PTHREAD_MUTEX_ERRORCHECK
argument, so we add macro OT_MUTEXINIT_NOCHECK to not check for
pthread_mutexattr_init return value.
It is set only when compiling with MSVC on Windows, but can also be set
when defining C++ flags.

In Log::Initialization() pthread_mutexattr_init was not called on Windows.
This looks strange, it is now called on all systems.

Note that several calls of pthread_mutexattr_init check for return value
without passing PTHREAD_MUTEX_ERRORCHECK, which does not make sense.