MSVC: Port I/O functions to MSVC
I/O functions are implemented in Os, Path and WrapperCommonFunctions.
They must be ported to MSVC (because of lack of dirname, mkdir, etc).
Moreover, we may want to get rid of wrapper stuff, which is deprecated.

Thus create MakeDirectory and DeleteDirectory methods into Os class, and
remove use of WrapperError from Path.cxx.