Translate Fortran into C with f2c
There is no Fortran compiler in Microsoft Visual Studio Express, it
is available only in the Professional edition.  It is thus much
easier to build OpenTURNS with Visual if there is no Fortran code.

Fortran files have not been deleted from the repository, we may
need to call f2c again in the future.

Based on a patch by Julien Schueller.