Implement CovarianceModel::computeCrossCovariance
This allows to compute the covariance interaction between X & Y samples.
or X/p where X is a Sample, p a Point
It was already implemented (but private) in KrigingResult for Kriging
variance purposes. Now we implement the method in the CovarianceModel
class as the service is mandatory for some applications such as HSIC

We also implement unit tests in CovarianceModel_std tests and reduce
time for continuous integration purposes