[hmat] Implement computeLogDeterminant
First add an isFactorized attribut to HMatrix, providing information
about factorization (if already done). This is because it is not
homogeneous in hmat-oss & a check is required for computing the log-det

Then use the hmat method to get the log determinant. The method is
more generic than LLt decomposition

In case of failure (more numerical issues rather than factorization)
compute manually & provide a SpecFunc value.